Making the level
Please note, I created a separate scene for the level. I moved the player over to the scene from where it originally was.

Future changes I am going to make here in a little bit.
(Change the radius for the "vision" to a cone)
Add a smaller radius for the enemy "hearing"