Dry fire sound and interact changes.
Interact button changed from E to F so that the player does not lean right every time interact is pressed. Code set up to play dry fire if shoot is attempted with zero ammo. Did not leave sound attached because it sounded awful. Also coded a rudimentary zoom or ADS but commented it out because I was unsatisfied with how it looked.